+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add: `input` optionally takes a description string
+
 ## [0.2.0] - 2024-05-10
 
 - Add `parameterized_testing-minitest-spec` to support minitest/spec